requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.35.0
google-cloud-vision>=3.8.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pillow>=10.0.0
requests>=2.32.0
python-multipart>=0.0.12
aiosqlite>=0.20.0
httpx>=0.27.0
jinja2
sendgrid>=6.11.0
easyocr>=1.7.0
openai>=1.0.0
